pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
requests = "^2.27.1"
firebase-admin = "^5.2.0"
Flask = "^2.0.3"
line-bot-sdk = "^2.1.0"
qrcode = "^7.3.1"
liffpy = "^0.0.1-alpha.3"
image = "^1.5.33"
pandas = "^1.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"